created admin permissions for pages
Created admin permissions for specific pages (register, users and edit/delete books)

+++ routes/livros.py
@@ -1,9 +1,11 @@
 import sqlite3
-from flask import flash, redirect, render_template, request, jsonify, url_for
+from flask import redirect, render_template, request, jsonify, url_for, flash
+from flask_login import current_user
 from . import bp
 from db.database import get_db
 from datetime import datetime
 import pytz
+import re
 
 now = datetime.now(pytz.timezone('America/Sao_Paulo')).strftime('%Y-%m-%d %H:%M:%S')
 
@@ -38,8 +40,6 @@ def get_livros():
     
     except sqlite3.Error as e:
         return jsonify({'error': str(e)}), 500
-    finally:
-        db.close()
 
 def add_livro():
     titulo = request.form.get('titulo')
@@ -58,25 +58,27 @@ def add_livro():
             (titulo, autor, genero, paginas, now)
         )
         db.commit()
-        flash('Livro adicionado com sucesso!', 'success')
-        return redirect(url_for('routes.handle_livros'))
+        return jsonify({'message': 'Livro adicionado com sucesso'})
     except sqlite3.Error as e:
-        flash(f'Erro ao adicionar livro: {str(e)}', 'danger')
-        return redirect(url_for('routes.create_livro'))
-    finally:
-        db.close()
+        return jsonify({'error': str(e)}), 500
 
 @bp.route('/livro/<int:livro_id>', methods=['DELETE', 'POST', 'GET', 'PUT'])
 def handle_livro(livro_id):
+    if getattr(current_user, 'cargo', '') != 'admin':
+        flash("Acesso negado: Apenas administradores podem acessar esta página.", "danger")
+        return redirect(url_for('routes.forbidden'))
     if request.method == 'GET':
         return get_livro(livro_id)
+    
     elif request.method == 'POST':
         if request.form.get('_method') == 'DELETE':
             return delete_livro(livro_id)
         elif request.form.get('_method') == 'PUT':
             return update_livro(livro_id)
+
     elif request.method == 'PUT':
         return update_livro(livro_id)
+    
     elif request.method == 'DELETE':
         return delete_livro(livro_id)
 
@@ -92,22 +94,21 @@ def get_livro(livro_id):
             return jsonify({'error': 'Livro não encontrado'}), 404
     except sqlite3.Error as e:
         return jsonify({'error': str(e)}), 500
-    finally:
-        db.close()
 
 def delete_livro(livro_id):
     try:
         db = get_db()
         cursor = db.cursor()
-        cursor.execute('DELETE FROM livros WHERE id = ?', (livro_id,))
-        db.commit()
-        flash('Livro excluir com sucesso!', 'success')
-        return redirect(url_for('routes.handle_livros'))
+        cursor.execute('SELECT * FROM livros WHERE id = ?', (livro_id,))
+        livro = cursor.fetchone()
+        if livro:
+            cursor.execute('DELETE FROM livros WHERE id = ?', (livro_id,))
+            db.commit()
+            return jsonify({'message': 'Livro excluído com sucesso'})
+        else:
+            return jsonify({'error': 'Livro não encontrado'})
     except sqlite3.Error as e:
-        flash(f'Erro ao adicionar livro: {str(e)}', 'danger')
-        return redirect(url_for('routes.create_livro'))
-    finally:
-        db.close()
+        return jsonify({'error': str(e)}), 500
 
 def update_livro(livro_id):
     titulo = request.form.get('titulo')
@@ -117,22 +118,24 @@ def update_livro(livro_id):
 
     if not titulo or not autor or not genero or not paginas:
         return jsonify({'error': 'Todos os campos são obrigatórios'}), 400
-    
+
     try:
         db = get_db()
         cursor = db.cursor()
+        cursor.execute('SELECT * FROM livros WHERE id = ?', (livro_id,))
+        livro = cursor.fetchone()
+
+        if not livro:
+            return jsonify({'error': 'Livro não encontrado'}), 404
+
         cursor.execute(
             'UPDATE livros SET titulo = ?, autor = ?, genero = ?, paginas = ?, modified = ? WHERE id = ?',
             (titulo, autor, genero, paginas, now, livro_id)
         )
         db.commit()
-        flash('Livro editado com sucesso!', 'success')
-        return redirect(url_for('routes.handle_livros'))
+        return jsonify({'message': 'Livro atualizado com sucesso'})
     except sqlite3.Error as e:
-        flash(f'Erro ao adicionar livro: {str(e)}', 'danger')
-        return redirect(url_for('routes.create_livro'))
-    finally:
-        db.close()
+        return jsonify({'error': str(e)}), 500
 
 @bp.route('/livros/create', methods=['GET'])
 def create_livro():
